Move behavior-related message methods to trait
